Rename common.models.ModelGenerator
