+++ api/src/test/java/marquez/MarquezAppIntegrationTest.java
@@ -2,9 +2,9 @@ package marquez;
 
 import static java.time.Instant.EPOCH;
 import static marquez.Generator.newTimestamp;
-import static marquez.common.models.ModelGenerator.newDescription;
-import static marquez.common.models.ModelGenerator.newJobName;
-import static marquez.common.models.ModelGenerator.newSchemaLocation;
+import static marquez.common.models.CommonModelGenerator.newDescription;
+import static marquez.common.models.CommonModelGenerator.newJobName;
+import static marquez.common.models.CommonModelGenerator.newSchemaLocation;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.google.common.collect.ImmutableList;
