+++ api/src/test/java/marquez/db/DbTestUtils.java
@@ -2,13 +2,13 @@ package marquez.db;
 
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static marquez.Generator.newTimestamp;
-import static marquez.common.models.ModelGenerator.newContext;
-import static marquez.common.models.ModelGenerator.newDatasetName;
-import static marquez.common.models.ModelGenerator.newDescription;
-import static marquez.common.models.ModelGenerator.newExternalId;
-import static marquez.common.models.ModelGenerator.newLocation;
-import static marquez.common.models.ModelGenerator.newNamespaceName;
-import static marquez.common.models.ModelGenerator.newOwnerName;
+import static marquez.common.models.CommonModelGenerator.newContext;
+import static marquez.common.models.CommonModelGenerator.newDatasetName;
+import static marquez.common.models.CommonModelGenerator.newDescription;
+import static marquez.common.models.CommonModelGenerator.newExternalId;
+import static marquez.common.models.CommonModelGenerator.newLocation;
+import static marquez.common.models.CommonModelGenerator.newNamespaceName;
+import static marquez.common.models.CommonModelGenerator.newOwnerName;
 import static marquez.db.models.DbModelGenerator.newRowUuid;
 import static marquez.service.models.ServiceModelGenerator.newDbTableMeta;
 import static marquez.service.models.ServiceModelGenerator.newDbTableMetaWith;
