+++ api/src/test/java/marquez/service/models/ServiceModelGenerator.java
@@ -2,19 +2,19 @@ package marquez.service.models;
 
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static java.time.temporal.ChronoUnit.HOURS;
-import static marquez.common.models.ModelGenerator.newContext;
-import static marquez.common.models.ModelGenerator.newDatasetId;
-import static marquez.common.models.ModelGenerator.newDatasetIdsWith;
-import static marquez.common.models.ModelGenerator.newDatasetName;
-import static marquez.common.models.ModelGenerator.newDescription;
-import static marquez.common.models.ModelGenerator.newFields;
-import static marquez.common.models.ModelGenerator.newJobType;
-import static marquez.common.models.ModelGenerator.newLocation;
-import static marquez.common.models.ModelGenerator.newNamespaceName;
-import static marquez.common.models.ModelGenerator.newOwnerName;
-import static marquez.common.models.ModelGenerator.newSourceName;
-import static marquez.common.models.ModelGenerator.newTagName;
-import static marquez.common.models.ModelGenerator.newTagNames;
+import static marquez.common.models.CommonModelGenerator.newContext;
+import static marquez.common.models.CommonModelGenerator.newDatasetId;
+import static marquez.common.models.CommonModelGenerator.newDatasetIdsWith;
+import static marquez.common.models.CommonModelGenerator.newDatasetName;
+import static marquez.common.models.CommonModelGenerator.newDescription;
+import static marquez.common.models.CommonModelGenerator.newFields;
+import static marquez.common.models.CommonModelGenerator.newJobType;
+import static marquez.common.models.CommonModelGenerator.newLocation;
+import static marquez.common.models.CommonModelGenerator.newNamespaceName;
+import static marquez.common.models.CommonModelGenerator.newOwnerName;
+import static marquez.common.models.CommonModelGenerator.newSourceName;
+import static marquez.common.models.CommonModelGenerator.newTagName;
+import static marquez.common.models.CommonModelGenerator.newTagNames;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
